+++ source/components/keypad/keypad.js
@@ -1,7 +1,7 @@
 /* global Metro */
 (function(Metro, $) {
     'use strict';
-    var Utils = Metro.utils;
+    //var Utils = Metro.utils;
     var KeypadDefaultConfig = {
         keypadDeferred: 0,
         label: "",
@@ -254,8 +254,11 @@
                     }
                 }
 
-                element.trigger('change');
-                Utils.exec(o.onChange, [that.value], element[0]);
+                that._fireEvent('change', {
+                    val: that.val
+                })
+                // element.trigger('change');
+                // Utils.exec(o.onChange, [that.value], element[0]);
 
                 e.preventDefault();
                 e.stopPropagation();
@@ -274,6 +277,11 @@
 
                 e.preventDefault();
                 e.stopPropagation();
+            });
+
+            element.on(Metro.events.inputchange, function(){
+                if (that.value !== this.value)
+                    that.value = this.value;
             });
 
             if (o.target !== null) {
@@ -314,13 +322,21 @@
         },
 
         val: function(v){
+            var element = this.element;
 
             if (typeof v === "undefined") {
                 return this.value;
             }
 
-            this.value = v;
-            this.element[0].tagName === "INPUT" ? this.element.val(v) : this.element.text(v);
+            this.value = ""+v;
+
+            if (element[0].tagName === "INPUT") {
+                element.val(v);
+                // set cursor to end position
+            } else {
+                element.text(v)
+            }
+
             return this;
         },