+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 + [x] Slider: fixed work vertical slider for touchable devices
 + [x] Colors: improve work with hsl, hsv
 + [x] Spinner: fixed input not a number value, value more or less than defined in options.
++ [x] Keypad: fixed internal value type mismatch.
 
 ### 4.4.2
 #### Metro 4